⚗️ replace old cache with cache from package
